refactor: extract player tracking logic to partial class and modernize UI
   - Moved player tracking and online list event handlers from MainWindow.xaml.cs to a new partial class MainWindow.Players.cs.
   - Implemented DataTemplate for TrackedPlayer in MainWindow.xaml to replace manual UI construction in code-behind.
   - Modernized player list styling using DynamicResource brushes and Wpf.Ui components for better theme integration.
   - Added IsOnline and PlayTimeStr properties to TrackedPlayer model to support data binding.
   - Enhanced tracked player UI with context menus and improved status indicators.
